Try to speed up some more, at the cost of a non-uniform sample
Sort keeps the items in order if they have the same key, so earlier neighbors
will have a higher chance of being selected.

Does the order of the `cell` values actually have the same problem?

Re-sorting neighbors is probably overkill if cells remain the same quite often,
but to find out whether they do/shoud, the previous question needs to be
answered first.